Get splitting up into implementation and instantiation files correct
