+++ Grid/qcd/action/fermion/instantiation/GparityWilsonImplD/CompactWilsonCloverFermionInstantiationGparityWilsonImplD.cc
@@ -1,0 +1,1 @@
+../CompactWilsonCloverFermionInstantiation.cc.master
