+++ Grid/qcd/action/fermion/instantiation/WilsonAdjImplD/CompactWilsonCloverFermionInstantiationWilsonAdjImplD.cc
@@ -1,0 +1,1 @@
+../CompactWilsonCloverFermionInstantiation.cc.master
