+++ Grid/qcd/action/fermion/instantiation/WilsonImplD/CompactWilsonCloverFermionInstantiationWilsonImplD.cc
@@ -1,0 +1,1 @@
+../CompactWilsonCloverFermionInstantiation.cc.master
