+++ Grid/qcd/action/fermion/instantiation/WilsonTwoIndexSymmetricImplD/CompactWilsonCloverFermionInstantiationWilsonTwoIndexSymmetricImplD.cc
@@ -1,0 +1,1 @@
+../CompactWilsonCloverFermionInstantiation.cc.master
